minor fixes

+++ scripts/scripts/vaccinations/src/vax/process/process.py
@@ -9,11 +9,13 @@ def process_location(df: pd.DataFrame) -> pd.DataFrame:
     # Default columns for second doses
     if "people_vaccinated" not in df:
         df = df.assign(people_vaccinated=pd.NA)
+        df.people_vaccinated = df.people_vaccinated.astype("Int64")
     if "people_fully_vaccinated" not in df:
         df = df.assign(people_fully_vaccinated=pd.NA)
+        df.people_fully_vaccinated = df.people_fully_vaccinated.astype("Int64")
     # Avoid decimals
     cols = ["total_vaccinations", "people_vaccinated", "people_fully_vaccinated"]
-    df[cols] = df[cols].astype("Int64").fillna(pd.NA)
+    df[cols] = df[cols].astype(float).astype("Int64").fillna(pd.NA)
     # Order columns and rows
     usecols = [
         "location", "date", "vaccine", "source_url", "total_vaccinations", "people_vaccinated",
